fix(events): add private router

+++ events/frontend/src/components/event/EventItem/EventItem.jsx
@@ -1,11 +1,20 @@
-import { Link, useSubmit } from 'react-router-dom';
+import { Link, useNavigate, useSubmit } from 'react-router-dom';
 
+import { getAuthToken } from '../../../utils/token';
 import classes from './EventItem.module.css';
 
 export function EventItem({ event }) {
+  const navigate = useNavigate();
   const submit = useSubmit();
 
   const startDeleteHandler = () => {
+    const token = getAuthToken();
+
+    if (!token) {
+      navigate('/auth/sign-in');
+      return;
+    }
+
     const proceed = window.confirm('Are you sure?');
     if (proceed) submit(null, { method: 'delete' });
   };
